* fix photos bug

+++ projects/app_photos/main/app/ui_event_handler.h
@@ -24,7 +24,7 @@ char *ui_get_delete_big_photo_path(void);
 bool ui_get_touch_show_right_big_photo_flag(void);
 bool ui_get_touch_show_left_big_photo_flag(void);
 bool ui_touch_is_video_image_flag(void);
-void ui_set_video_bar_s(uint32_t curr_s, uint32_t total_s);
+void ui_set_video_bar_s(double curr_s, double total_s);
 bool ui_get_touch_video_bar_press_flag(void);
 bool ui_get_touch_video_bar_release_flag(void);
 double ui_get_video_bar_value(void);
